search: move progress to graphqlbackend
As we improve on progress, we will need more access to "search
internals" which currently live in the graphqlbackend. I wanted to avoid
this, but for practicality purposes this is best.
